Functions

TreeKnit.computeMCCs

TreeKnit.naive_mccs

TreeKnit.resolve!

TreeKnit.runopt

TreeKnit.treeknit

Function used for the CLI

TreeKnit.treeknit — Function

treeknit

Arguments

nwk1: Newick file for first tree

nwk2: Newick file for second tree

Options

-o, --outdir <arg>: directory for results; Example -o=treeknit_results

-g, --gamma <arg>: value of γ; Example -g=2

--seq-lengths <arg>: length of the sequences. Example: --seq-length "1500 2000"

--n-mcmc-it <arg>: number of MCMC iterations per leaf; default 25

Flags

--naive: Naive inference (overrides -g).

--no-likelihood: Do not use branch length likelihood test to sort between different MCCs

--no-resolve: Do not attempt to resolve trees before inferring MCCs.

-v, --verbose: verbosity

Computing Maximal Compatible Clades (MCCs) for pairs

Main functions

TreeKnit.computeMCCs — Function

computeMCCs(
	t1::Tree, t2::Tree, oa::OptArgs = OptArgs();
	naive = false, seqlengths = [1,1],
)

Compute pairwise MCCs for trees. Return MCCs and resolved splits. The computeMCCs! version resolves the input trees with newly found splits.

Inputs

oa::OptArgs

Controls parameters of the MCC inference (unless naive=true). See ?OptArgs for details.

In general, this should be set to true if more than two trees are used, and to false for only two trees (for speed).

naive = false

If true, use a naive estimation for MCCs, i.e. find all clades that have an exactly matching topology in all trees.

Else, use a pseudo-parsimonious method based (mostly) on topology. The method runopt(oa,t1,t2) is called on every pair of trees.

For pairs of trees only

TreeKnit.naive_mccs — Function

naive_mccs(treelist)

Find sets of nodes which are:

clades in all trees of treelist,

all subclades of nodes are clades in all trees of treelist (both of these properties define consistency),

maximal: adding a node to a set results it in not being a clade in at least one of the trees.

All the trees of treelist should share the same leaf nodes.

TreeKnit.runopt — Function

runopt(t1::Tree, t2::Tree; kwargs...)
	runopt(oa::OptArgs, t1::Tree, t2::Tree)
	runopt(oa::OptArgs, trees::Dict{<:Any,<:Tree})

Run optimization at constant γ. See ?Optargs for arguments. In the first form, keyword arguments are given to OptArgs.

Resolving trees

Using topology

TreeKnit.resolve! — Method

resolve!(t1::Tree, t2::Tree; tau=0.)

Resolve t1 using splits of t2 and inversely. Every split of t2 a tree that is compatible with t1 is introduced in t1 with branch length tau (and inversely). Return new splits in each tree.