The OptArgs object

computeMCCs and several other functions take an OptArgs object that controls several parameters of the inference process.

TreeKnit.OptArgs — Type

struct OptArgs

Storing parameters for SplitGraph.runopt function.

General

γ::Real = 2

itmax::Int64 = 15: maximal number of iterations of naive MCCs / SA cycles.

likelihood_sort::Bool = true: sort equivalent configurations using likelihood test based on branch length.

resolve::Bool = true: try to resolve trees while finding MCCs.

seq_lengths: lengths of sequences that trees were built from. Used in likelihood calculations. This is initialized from other input arguments, and defaults to sequences of length one.

Simulated annealing

nMCMC::Int = 25: The number of MCMC iterations for a tree of n leaves is nMCMC*n.

cooling_schedule = :geometric: type of cooling schedule (:geometric, :linear, :acos)

Tmin::Float64 = 0.05: minimal temperature of SA.

Tmax::Float64 = 0.8: maximal temperature of SA.

nT::Int = 3000: number of steps in the cooling schedule

Verbosity

verbose::Bool=false: first level of verbosity

vv::Bool = false: second level of verbosity